requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx>=0.24.0,<0.25.0